requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlmodel==0.0.14
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.5
aiofiles==23.2.1
Pillow==10.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
alembic==1.13.1
aiosmtplib==3.0.1
sqladmin==0.16.0
itsdangerous==2.1.2
slowapi==0.1.9
